Expanded Demo
The demo has been built up into four sections. DemoTutorial is intended to show the player basic platforming and Birdsong mechanics. DemoHub has a treasure chest and three keys that "unlock" it (Doing so currently does nothing, however.) DemoEast has an NPC that wanders back and forth to be cured via Birdsong. (See notes below on Back and Forth movement.) DemoWest currently only has an entrance/exit.

Additionally, in the course of building DemoEast, we created the Smart Back and Forth extension, which has some problems that I'm not entirely clear on as of yet. Currently we're using Timed Back and Forth instead, but I would like to finish it nontheless.