Keybindings from storage!
We've managed to get the Storage feature to do what we needed! We can save and load keybinds this way, which I was able to test by manually editing the Storage object. I will still, however, need to build a menu for *setting* those keybinds, and the SaveKeybinds event may want some added failsafes as well, I think.